feat-routing

+++ MusicaListFront/src/app/admin-genres/admin-genres.component.ts
@@ -4,18 +4,20 @@ import {GeneroMusicalService} from "../services/generoMusical.service";
 import {Router} from "@angular/router";
 import {Cancion} from "../models/cancion.model";
 import {CancionService} from "../services/cancion.service";
+import {AdministradorService} from "../services/administrador.service";
 
 @Component({
   selector: 'app-admin-genres',
   templateUrl: './admin-genres.component.html',
   styleUrls: ['./admin-genres.component.css']
 })
-export class AdminGenresComponent implements OnInit{
+export class AdminGenresComponent implements OnInit {
   generosMusicales: GeneroMusical[] = [];
   canciones: Cancion[] = [];
 
-  constructor(private generoMusicalService : GeneroMusicalService) {
+  constructor(private generoMusicalService: GeneroMusicalService, private administradorService: AdministradorService) {
   }
+
   ngOnInit() {
     this.generoMusicalService.getAllGenerosMusicales().subscribe(
       (generosMusicales: GeneroMusical[]) => {
@@ -26,8 +28,12 @@ export class AdminGenresComponent implements OnInit{
 
 
   eliminarGenero(id: number) {
-    this.generoMusicalService.deleteGeneroMusicalById(id).subscribe((genero: GeneroMusical) => {
-      this.generosMusicales = this.generosMusicales.filter((genero: GeneroMusical) => genero.idGenero != id)
-    })
+    if (this.administradorService.administrador.idCuenta == -1)
+      alert("Es necesario iniciar sesion para borrar un genero")
+    else {
+      this.generoMusicalService.deleteGeneroMusicalById(id).subscribe((genero: GeneroMusical) => {
+        this.generosMusicales = this.generosMusicales.filter((genero: GeneroMusical) => genero.idGenero != id)
+      })
+    }
   }
 }
